requirements.txt
pandas
numpy
FinanceDataReader
yfinance
python-telegram-bot
requests
google-genai
exchange_calendars
python-dotenv
